+++ CHANGELOG.md
@@ -19,6 +19,9 @@ Change categories:
 
 ### Added
 
+- firmware_update: Add GetDeviceMetaData command
+  - Add `encode_get_device_meta_data_req()`
+  - Add `decode_get_device_meta_data_resp()`
 - firmware_update: Add GetMetaData command
   - Add `encode_get_meta_data_resp()`
   - Add `decode_get_meta_data_req()`